Extract turbo addon and remove button mask duplicate
